+++ crmeb-service/src/main/java/com/cshy/service/dao/system/SysDictDataMapper.java
@@ -42,7 +42,7 @@ public interface SysDictDataMapper
      * @param dictCode 字典数据ID
      * @return 字典数据
      */
-     SysDictData selectDictDataById(Long dictCode);
+     SysDictData selectDictDataById(Integer dictCode);
 
     /**
      * 查询字典数据
@@ -58,7 +58,7 @@ public interface SysDictDataMapper
      * @param dictCode 字典数据ID
      * @return 结果
      */
-     int deleteDictDataById(Long dictCode);
+     int deleteDictDataById(Integer dictCode);
 
     /**
      * 批量删除字典数据信息
@@ -66,7 +66,7 @@ public interface SysDictDataMapper
      * @param dictCodes 需要删除的字典数据ID
      * @return 结果
      */
-     int deleteDictDataByIds(Long[] dictCodes);
+     int deleteDictDataByIds(Integer[] dictCodes);
 
     /**
      * 新增字典数据信息
